+++ src/lib/external/index.ts
@@ -1,2 +1,13 @@
+/**
+ * External documentation functionality
+ * Re-exports all external-doc related functions and types
+ */
+
+// Re-export types
+export type * from "./types"
+
+// Re-export fetching functions
 export * from "./fetch"
+
+// Re-export policy functions and errors
 export * from "./policy"
